+++ app/models/Switch.scala
@@ -12,8 +12,8 @@ case class Switch(
   portsNumber: Option[Int],
   buildShortName: Option[String],
   floorNumber: Option[Int],
-  positionTop: Option[Int],
-  positionLeft: Option[Int],
+  positionTop: Option[Float],
+  positionLeft: Option[Float],
   upSwitchName: Option[String],
   upSwitchMAC: Option[String],
   upLink: Option[String]
